+++ lib/api/place_api_provider.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 import 'package:flutter/material.dart';
+import 'package:google_maps_places_autocomplete_widgets/api/autocomplete_types.dart';
 import 'package:http/http.dart';
 
 import '/model/place.dart';
@@ -103,13 +104,24 @@ result["predictions"] =
   ///   instead of address type information.
   Future<List<Suggestion>> fetchSuggestions(String input,
       {bool includeFullSuggestionDetails = false,
-      bool postalCodeLookup = false}) async {
+      required List<AutoCompleteType> types}) async {
+
+    /// OK, we need to check the types array..
+    String typesString = '';
+    for(final type in types) {
+      if(type.onlySingleValueAllowed && types.length>1) {
+        throw Exception('If $type is specified then it is the ONLY autocomplete type allowed by Google Places. See https://developers.google.com/maps/documentation/places/web-service/autocomplete#types');
+      }
+      if(typesString.length>0) typesString += '|';
+      typesString += type.typeString;
+    }
+    if(types.length>5) {
+        throw Exception('A maximum of 5 autocomplete types are allowed by Google Places. See https://developers.google.com/maps/documentation/places/web-service/autocomplete#types');
+    }
+
     final Map<String, dynamic> parameters = <String, dynamic>{
       'input': input,
-      'types': postalCodeLookup
-          ? 'postal_code'
-          : 'address', // this is for looking up fully qualified addresses
-      // Could be used for ZIP lookups//   'types': 'postal_code',
+      'types': typesString,
       'key': mapsApiKey,
       'sessiontoken': sessionToken
     };