+++ lib/widgets/address_autocomplete_textformfield.dart
@@ -4,6 +4,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:google_maps_places_autocomplete_widgets/api/autocomplete_types.dart';
 import 'package:uuid/uuid.dart';
 
 import 'address_autocomplete_generic.dart';
@@ -94,9 +95,23 @@ class AddressAutocompleteTextFormField
   @override
   final String? language;
 
-  ///PostalCode lookup instead of address lookup (defaults to false)
+  /// (deprecated) PostalCode lookup instead of address lookup (defaults to false)
+  /// (This has now been deprecated and replaced with [type] parameter `type:AutoCompleteType.postalCode`).
+  @Deprecated("If passing true use `type:AutoCompleteType.postalCode` instead. (false == passing `type:AutoCompleteType.address`")
+  final bool? postalCodeLookup;
+
+  /// Single AutoCompleteType enum for type of information to autocomplete
+  /// (Defaults to [AutoCompleteType.address] , use [AutoCompleteType.postalCode] for Zipcode lookup)
+  /// This must be null if [types] is supplied
   @override
-  final bool postalCodeLookup;
+  final AutoCompleteType? type;
+
+  /// Can be used as an *alternative* to [type] and allows a list of up to 5 AutoCompleteType enums
+  /// to specify the types of autocomplete information to lookup.
+  /// (Google places api only allows 5 types maximum)
+  /// ([type] must not be supplied if [types] is used).
+  @override
+  final List<AutoCompleteType>? types;
 
   ///debounce time in milliseconds (default 600)
   @override
@@ -232,7 +247,9 @@ class AddressAutocompleteTextFormField
     this.elevation,
     this.overlayOffset = 4,
     this.showGoogleTradeMark = false,
-    this.postalCodeLookup = false,
+    this.postalCodeLookup,
+    this.type,
+    this.types,
     this.componentCountry,
     this.language,
 
@@ -282,7 +299,11 @@ class AddressAutocompleteTextFormField
     this.enableIMEPersonalizedLearning = true,
     this.mouseCursor,
     this.contextMenuBuilder,
-  });
+  }) : assert( (postalCodeLookup==true && type==null && types==null) 
+              || (postalCodeLookup==false && type==null && types==null)
+              || (postalCodeLookup==null && type==null && types==null)
+              || (postalCodeLookup==null && type!=null && types==null)
+              || (postalCodeLookup==null && type==null && types!=null), 'You can only supply value for [type], [types] (or deprecated `postalCodeLookup`).  No combinations allowed');
 
   @override
   State<StatefulWidget> createState() =>